Add DinoPark Packs routes
This adds routes for DinoPark Packs described in the hosted swagger file as well as routes present in the `sudo` app source code that aren't present in the swagger file.

Some of the descriptions of the `sudo` routes are missing as I couldn't determine by reading the code what they did.

I haven't marked any of the routes as internal or public as I'm unsure how to determine that.